# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "INVENTRAK Inventory Management API",
    "description": "Backend API for the INVENTRAK inventory management system. Both the SQLite (Express) backend and the npm-free fallback expose this identical contract. Live docs: GET /api/docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local backend"
    }
  ],
...
  "paths": {
...
    "/api/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "operationId": "listCustomers",
        "summary": "Customer Records with per-customer aggregates (admin only)",
        "description": "The business entity behind orders and sales, distinct from accounts. Admin-only: it exposes what one person ordered and what they spent. Each row carries inquiry_count, total_spend and last_activity.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Customer list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "operationId": "getCustomer",
        "summary": "One customer's record with their order and purchase history (admin only)",
        "description": "The query the old schema could not answer: customer_name was free text repeated on every row, so a single person's history across orders was unretrievable.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer detail",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Customer"
                    }
                  ],
                  "type": "object",
                  "properties": {
                    "inquiries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "status": {
                            "type": "string"
                          },
                          "estimated_cost": {
                            "type": "number"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "status_history": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "sales": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "product_id": {
                            "type": "integer"
                          },
                          "qty": {
                            "type": "number"
                          },
                          "total_amount": {
                            "type": "number"
                          },
                          "transaction_date": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Customers"
        ],
        "operationId": "updateCustomer",
        "summary": "Edit a Customer Record (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Another customer already uses this email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Admin access required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Ok": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "Customer": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "business_name": {
            "type": "string",
            "nullable": true
          },
          "contact_number": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true,
            "description": "Unique when present (case-insensitive). The most stable identifier a walk-in customer gives."
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "user_id": {
            "type": "integer",
            "nullable": true,
            "description": "Optional link to the account this customer signs in with; null for walk-ins, since guest checkout is first-class."
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "inquiry_count": {
            "type": "integer",
            "description": "How many order inquiries belong to this customer."
          },
          "total_spend": {
            "type": "number",
            "description": "Sum of their linked sales' total_amount."
          },
          "last_activity": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CustomerInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "business_name": {
            "type": "string",
            "maxLength": 150
          },
          "contact_number": {
            "type": "string",
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          },
          "address": {
            "type": "string",
            "maxLength": 300
          }
        }
      },
...
    }
  }
}